// packages/vincent-rest-api/src/common/api/models/ICreateAppVersionDef.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */

export type ICreateAppVersionDef = {
  /**
   * List of tool identities to include in this version
   */
  tools: Array<string>;
  /**
   * Changelog information for this version
   */
  changes: string;
};